+++ docs/source/content/changelog.rst
@@ -5,6 +5,7 @@ v0.1.5
 ^^^^^^^^^^^^^^^^^^
 * Bug fixes
 * Now the file name of ``Autosaver.save`` is optional, will use ``datetime.datetime.now().strftime(r"%m%d-%H%M%S") + ".csv"`` if left empty
+* Add ``__recompute__`` functionality for ``checkpoint``
 
 v0.1.4
 ^^^^^^^^^^^^^^^^